src/ai_server/requirements.txt
# Puython 3.10 suggested

numpy
fastapi
grpcio-tools
grpcio
python-dotenv
fastembed-gpu
nvidia-cudnn-cu12
uvicorn
grpcio-reflection
argparse
--extra-index-url https://download.pytorch.org/whl/cu124
torch
torchvision
torchaudio
opentelemetry-api
opentelemetry-sdk
pyyaml
cachetools
vllm